separate node and container code from tree
